Backend
Worked further on the backend, changed the way with how keybinds will
be initialized, dedicated a file to storing the table with the command
functions. These will have no inputs and no outputs. Indexes
within hashtables are also handled with a macro instead of a dedicated
function, there is a global general hash function in util.c.