Buggy key sequences broken UI
The key sequences are working, but only on the first layer. A
segmentation fault occurs when a secondary layer should be added. The
start menu UI is still broken, but this is a matter of interpreting the
BMP format. Currently it seems that the logo is upside down and flipped.