Useful functions
Implemented useful functions: up, down, left, and right. These tell
the currently active screen, through the use of a signal or "local"
function that the user requests the cursor to go up, down, left, or
right. This will be used later for the menu selection. Cleaned up some
of the rendering code in start.c and fixed a very small clang warning.