Working colors
The colors are now working. Had a few issues with buffer hygiene in
configuration.c and DEBUG_MSG defining the variable 'i' causing weird
errors. Those are fixed. Created a structure to house colors that hold a
bit that states whether the color is present or not (only present colors
are registered to make the application have quicker start up times).