Start.c refactor
Refactored the code which renders the start screen. It was
quite messy. It still is a little bit, but I feel that it is more
readable. There is a caveat, a for loop is run every frame to measure
the length of the longest element in the right table menu. This should
only be done once.