Further testing and line wrapping
Did some further testing and wrote fib.s using assembled.
It has been handy for discovering more things to do with assembled
and one which I would find somewhat nice is line wrapping. I have had a
new idea of how to implement it since last time. Wrapping works somewhat
on the y-axis, we need to wrap on the x-axis and ensure that other
columns also have the same y-distortion when drawing the contents.