Dynamic load menu
Changed the code of the file load screen's render function to display
the directory listing box more dynamically. I have also encountered a
weird bug (if you could call it that). Essentially the code for
Cltr + Left Arrow in the kitty terminal is different than the code for
it in VSCode's terminal.