Improved files
I have improved the file switching, although there is still a weird
problem with it. This problem basically doesn't allow it to access all
open text files when switching to the right. The file load screen has
been made so that it stops displaying the "(null)" string. fpath2abs
has been secured against NULL pointers. The interface is now also able
to keep track of the last screen the user was on, allowing for an
escape mechanism from the load screen (or any other screen).