Optimizations
Optimized for CPU utilization. Currently the CPU usage stays mostly at
0% while editing / loading files, spikes to < 1% when typing and
editing. And is much higher when on the start screen. This was
accomplished through the use of a boolean stating whether there is
input from the user. If there is no input from the user, and the
screen has the option SCR_OPT_ON_UPDATE enabled, then only an 8
millisecond delay will be done. If the screen has SCR_OPT_ALWAYS
then the screen will be drawn on every iteration.