Line selection
Got to work on line selection. It is still extremely finicky and needs a
lot more testing; however, it is working to a degree. There are lots of
left over bits of code that I need to clean up eventually. I also find
the current codebase a bit messy, so it would be of use to rename and
introduce some sort of naming convention.