Tidying of selections and line movement
I made the selections on appear on the currently active buffer. Another
thing to possibly do would be to save the current selection when the
user switches off a buffer, and then restore the end point back into the
cursor x when the user returns to the current buffer (if selection is
enabled). Started writing the functions which will move lines up and
down.