Added line movement
Implemented signle line movement. Next will be multiple line movement
and moving lines across every buffer. There is also a problem with the
editor screen where it falls out of desync if a line is moved down and
then the cursor is moved to the very bottom of the screen. Moving the
cursor to the top somehow resynchronizes the cursors and the the user
can go to the bottom of the screen again without offsetting the editor.