Multibuffer line movement
Worked more on line movement. There is some desynchrnoziation between
where the cursor is and the current_element in each buffer when moving
lines. I worked on fixing this, but there are still many ways to
desynchronize them. Multiline multibuffer movement does not work;
however, a single line works somewhat. More work needs to be put into
keeping the buffers and cursor synchronized.

A segmentation fault also occurs when moving a single line up from the
bottom of the file.