Convention
Implemented a naming convention to make the source code easier to read.
This involved lots of renaming and such. Removed many global variables,
now there are only two. The removal of the global variables resulted in
include errors, however, those are resolved for now. I am going to work
on cleaning up the includes.