Local codes simplified, fixed some bugs
I simplified the local codes so now the values argument is being
utilized more allowing for less cases to be in each switch statement. I
have fixed the issue with buffer_move_ln_up resulting in a segmentation
fault when used at the bottom of a file, and in some places I forgot to
implement the AS_Ctx structure, and was still using the old global
variables.