Syntax highlighting architecture
Built out an architecture for syntax highlighting. There will
essentially be a series of backends from which one is chose. The chosen
backend is then given all of the lines in the file and it computes a
linked list of structures. These structures define sections of the line
which should be colored using a defined color.