Fixed asm.c, beginning to color
There were a few logical errors in asm.c, these have been fixed and it
is now outputting proper values. I started work on the function which
will be used to display the syntactically highlighted characters. So far
everything is black for some reason. I am presuming that the colors are
not defined, even though they are.