New errors, figured out no openeing error
There are some new errors with line deletion and line creation.
Somewhere the links are being screwed up, I am not sure where as up to
this point this has not happened. The no opening error was caused by the
default case not checking the first character to see if it was valid,
this put it into an infinite loop causing it to eventually allocate all
memory in the system and crashing.