Backends, visual bugs
The syntax.c file can now determine the proper backend to use based on
the file extension, it is not the most efficient, but for now it gets
the job done. A visual bug has been fixed where there would essentially
be a ghost layer of text (forgot to remove a mvprintw). A new visual bug
has also been discovered which makes it impossible to see the last few
lines of a text file if said lines are wrapped.